Cleanups and a Windows build fix
