+++ src/tests/test_mac.cpp
@@ -70,7 +70,7 @@ class Message_Auth_Tests : public Text_Based_Test
          }
    };
 
-      BOTAN_REGISTER_TEST("mac", Message_Auth_Tests);
+BOTAN_REGISTER_TEST("mac", Message_Auth_Tests);
 
 #endif
